Move is_power_of_2
