+++ include/boost/simd/meta/is_power_of_2.hpp
